pyproject.toml
[project]
name = "autonomous-job-finder-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "langchain-google-genai",
    "langgraph",
    "langchain-core",
    "pydantic",
    "requests",
    "python-dotenv",
    "pyyaml",
]
